+++ lib/doorkeeper/models/access_token_mixin.rb
@@ -95,7 +95,7 @@ module Doorkeeper
                               resource_owner_or_id
                             end
         token = last_authorized_token_for(application.try(:id), resource_owner_id)
-        if token && scopes_match?(token.scopes, scopes, application.try(:scopes))
+        if token && scopes_match_exactly?(token.scopes, scopes, application.try(:scopes))
           token
         end
       end
@@ -116,6 +116,28 @@ module Doorkeeper
       def scopes_match?(token_scopes, param_scopes, app_scopes)
         (!token_scopes.present? && !param_scopes.present?) ||
           Doorkeeper::OAuth::Helpers::ScopeChecker.match?(
+            token_scopes.to_s,
+            param_scopes,
+            app_scopes
+          )
+      end
+
+      # Checks whether the token scopes match exactly the scopes from the parameters or
+      # Application scopes (if present).
+      #
+      # @param token_scopes [#to_s]
+      #   set of scopes (any object that responds to `#to_s`)
+      # @param param_scopes [String]
+      #   scopes from params
+      # @param app_scopes [String]
+      #   Application scopes
+      #
+      # @return [Boolean] true if all scopes are blank or exact matches
+      #   and false in other cases
+      #
+      def scopes_match_exactly?(token_scopes, param_scopes, app_scopes)
+        (!token_scopes.present? && !param_scopes.present?) ||
+          Doorkeeper::OAuth::Helpers::ScopeChecker.match_exactly?(
             token_scopes.to_s,
             param_scopes,
             app_scopes